+++ CHANGES.rst
@@ -149,6 +149,12 @@ Bugs fixed
 * #12995: Significantly improve performance when building the search index
   for Chinese languages.
   Patch by Adam Turner.
+* #12767: :py:meth:`.Builder.write` is typed as ``final``, meaning that the
+  :event:`write-started` event may be relied upon by extensions.
+  A new :py:meth:`.Builder.write_documents` method has been added to
+  control how documents are written.
+  This is intended for builders that do not output a file for each document.
+  Patch by Adam Turner.
 
 
 Testing